Custom icons list is only meant to be for icons where you can change the icon colour to custom
