+++ src/components/Icon/index.js
@@ -1,11 +1,27 @@
 import React, {PureComponent} from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
-import CUSTOM_ICONS from './custom-icons'
 
 const ICON_PATH = process.env.ICON_PATH || (typeof (window) !== 'undefined' && window['ICON_PATH']) || '/icons.svg'
 const SIZES = ['small', 'medium', 'large']
 const COLORS = ['white', 'typegrey', 'inputgrey', 'typecyan', 'custom']
+
+const CUSTOM_ICONS = [
+  'cross',
+  'facebook',
+  'facebook-brand',
+  'google',
+  'google-brand',
+  'renewable',
+  'star-half',
+  'star',
+  'starline-half',
+  'starline',
+  'tick',
+  'twitter',
+  'twitter-brand',
+  'uswitch'
+]
 
 export default class Icon extends PureComponent {
   get realColor () {
